\name{whittermore}

\alias{whittermore}

\title{Whittermore's statistic}

\description{
Whittermore's statistic is defined as follows:


\deqn{W=\frac{n-1}{n}r^{'} D r}{W = ((n-1) / n ) r' D r}

where \eqn{r^{'}=[O_1/O_+, \ldots, O_n/O_+]}{ r' = [O_1/O_+, \ldots, O_n/O_+]}
and \eqn{D}{D} is a distance matrix between the centroids of the areas.

It can be used to assess whether the region under study tends to cluster
or not.
}

\seealso{
DCluster, whittermore.stat, whittermore.boot, whittermore.pboot
}

\references{
Whittermore, A. S. and Friend, N. and Byron, W. and Brown, J. R. and Holly, E. A. (1987). A test to detect clusters of disease. Biometrika 74, 631-635.
}


\keyword{spatial}
